Improved Ecosia's TopSites selection
