+++ Storage/DefaultSuggestedSites.swift
@@ -3,8 +3,11 @@
 // file, You can obtain one at http://mozilla.org/MPL/2.0/
 
 import Foundation
+// Ecosia: Core import
+import Core
 
 open class DefaultSuggestedSites {
+    /* Ecosia: Replace Default Suggested Sites
     public static let urlMap = [
         "https://www.amazon.com/": [
             "as": "https://www.amazon.in",
@@ -129,6 +132,35 @@ open class DefaultSuggestedSites {
                 faviconUrl: "asset://jdLogo",
                 trackingId: 705,
                 title: "京东"
+            )
+        ]
+    ]
+     */
+    public static let sites = [
+        "default": [
+            SuggestedSiteData(
+                url: Environment.current.urlProvider.financialReports.absoluteString,
+                bgColor: "0x000000",
+                imageUrl: "asset://suggestedsites_ecosia-org",
+                faviconUrl: "asset://defaultFavicon",
+                trackingId: 0,
+                title: NSLocalizedString("Financial reports", tableName: "Ecosia", comment: "")
+            ),
+            SuggestedSiteData(
+                url: Environment.current.urlProvider.privacy.absoluteString,
+                bgColor: "0x000000",
+                imageUrl: "asset://suggestedsites_ecosia-org",
+                faviconUrl: "asset://defaultFavicon",
+                trackingId: 0,
+                title: NSLocalizedString("Privacy", tableName: "Ecosia", comment: "")
+            ),
+            SuggestedSiteData(
+                url: Environment.current.urlProvider.blog.absoluteString,
+                bgColor: "0x000000",
+                imageUrl: "asset://suggestedsites_ecosia-org",
+                faviconUrl: "asset://defaultFavicon",
+                trackingId: 0,
+                title: NSLocalizedString("Trees update", tableName: "Ecosia", comment: "")
             )
         ]
     ]
